feat: fetch data directly in the inngest function
