--- apps/web/src/app/actions/fii/schemas.ts
@@ -1,21 +1,0 @@
-export interface FiiData {
-  actualValue: number;
-  dividendYield: number;
-  name: string;
-  pvp: number;
-  segment?: string;
-  yield: {
-    lastYield?: {
-      value: number;
-      percentage: number;
-      basePrice: number;
-      date: string;
-    };
-    nextYield?: {
-      value: number;
-      percentage: number;
-      basePrice: number;
-      date: string;
-    };
-  };
-}
