+++ apps/web/src/models/notion.ts
@@ -17,6 +17,28 @@ export interface UpdateNotionDataSourceFiMessage {
   timeZone?: TimeZone;
 }
 
+export interface FiiData {
+  actualValue: number;
+  dividendYield: number;
+  name: string;
+  pvp: number;
+  segment?: string;
+  yield: {
+    lastYield?: {
+      value: number;
+      percentage: number;
+      basePrice: number;
+      date: string;
+    };
+    nextYield?: {
+      value: number;
+      percentage: number;
+      basePrice: number;
+      date: string;
+    };
+  };
+}
+
 export interface NotionReducePropertiesOptions {
   timeZone?: TimeZone;
 }
